refactor: 등록 및 수정 페이지 스키마 분리
