+++ src/constants/postForm.ts
@@ -1,5 +1,3 @@
-import { z } from 'zod';
-
 import { POST_TYPES } from './postTypes';
 
 export const PRICE_RADIO_LIST = [
@@ -7,44 +5,6 @@ export const PRICE_RADIO_LIST = [
   { id: POST_TYPES.online, label: POST_TYPES.online, value: 1 },
   { id: POST_TYPES['clan-recruitment'], label: POST_TYPES['clan-recruitment'], value: 2 },
   { id: POST_TYPES['game-strategy'], label: POST_TYPES['game-strategy'], value: 3 },
-];
-
-const POST_SCHEMA_DEFAULT = z.object({
-  price: z.string(),
-  title: z.string().min(1),
-  description: z.string().min(5),
-});
-
-export const PostSchema = [
-  POST_SCHEMA_DEFAULT.extend({
-    headcount: z.string().min(1),
-    address: z.string().min(1),
-    discord: z.string().min(1),
-    date: z.string().min(1),
-    startTime: z.string().min(1),
-    endTime: z.string().min(1),
-    bannerImageUrl: z.string().optional(),
-    subImageUrls: z.string().optional(),
-  }),
-
-  POST_SCHEMA_DEFAULT.extend({
-    headcount: z.string().min(1),
-    discord: z.string().min(1),
-    date: z.string().min(1),
-    startTime: z.string().min(1),
-    endTime: z.string().min(1),
-  }),
-
-  POST_SCHEMA_DEFAULT.extend({
-    discord: z.string().min(1),
-    bannerImageUrl: z.string().optional(),
-    subImageUrls: z.string().optional(),
-  }),
-
-  POST_SCHEMA_DEFAULT.extend({
-    bannerImageUrl: z.string().optional(),
-    subImageUrls: z.string().optional(),
-  }),
 ];
 
 export const SCRIPT_URL = '//t1.daumcdn.net/mapjsapi/bundle/postcode/prod/postcode.v2.js';
